[Refactoring] fix review comments
